Change the type definition of the `CompactTime` to be embedded `*time.Time`
Before this commit, `CompactTime` was a type alias of `time.Time`;
That is not a reference type so `CompactTime` cannot represent the nil.
Thus probably it will be always entered [this](https://github.com/brocaar/lora-gateway-bridge/blob/359742accabc2c5c3e9cd6987cb21f7597330f3e/internal/gateway/backend.go#L447) `if` clause.

This commit changes the definition of that to make it possible to
represent the nil value.