Care an empty string value that is contained in RXPK as `time` property
Some vendor's gateway sends an empty string value as `time` property of
Rx packet (an empty string means "ignore me", this is terrible...)
In the current implementation, the gateway-bridge raises errors at
[here](https://github.com/brocaar/lora-gateway-bridge/blob/359742accabc2c5c3e9cd6987cb21f7597330f3e/internal/gateway/structs.go#L324)
when an empty `time` has come so it cannot receive any packets if
`time` is an empty string.

This patch cares this problem.