Add support for unlimited channels and radios.
This commit removes the restriction of 2 radios and 8 channels and allows any number of
channels (and subsequently any number of radios) to be be configured by the gateway bridge.
The code automatically adds and configures the necessary number of radios required in order
to support the channels.

As was previously the case, the global_conf.json file needs to already contain entries
for the radios and channels in order for the new configuration to be successfully merged.
The limit on one Lora_std channel and one FSK channel remains.