Refactor gateway subscribe / unsubscribe.
Since the connect and disconnect channels were separate this could
result in a potential race condition in case of connect, disconnect,
connect (e.g. executed in the order connect, connect, disconnect).
Normally a quick connect, disconnect and connect would not happen.

This refactor changes the signaling into a single channel, to make sure
that all events are executed (and completed) in order.