Remove old packet-forwarder reconfiguration code.
Because there are many variations in configuration templates, this code
never worked on all gateways and confused more people than that it
offered a solution to. See
https://forum.chirpstack.io/t/who-is-using-the-gateway-profile-how-are-you-using-it/5091
for the discussion.

Note that the ChirpStack Gateway Bridge will continue to support the
re-configuration of gateways using the ChirpStack Concentratord. In this
case, the configuration is interpreted and handled by Concentratord.