Make sure uint32 token is truncated to uint16.
ChirpStack v3 only uses the uint16 range, but this makes it possible to
use the full uint32 range, while staying compatible with the Semtech UDP
Packet Forwarder protocol.

We truncate the uint32 token to uint16 and use this as the cache-key to
store the downlink frame. On ack, we retrieve the downlink frame using
the uint16 token and are able to recover the full uint32 token.